+++ moirai-parallel/src/lib.rs
@@ -93,7 +93,7 @@ pub use ops::{
     for_each_chunk_pair_mut_enumerated_with, for_each_chunk_quad_mut_enumerated_with,
     for_each_chunk_triple_mut_enumerated_with, for_each_index_with, for_each_mut_with,
     for_each_with, join, join_with, map_collect_index_with, map_collect_mut_with, map_collect_with,
-    map_reduce_with, reduce_index_with,
+    map_reduce_with, reduce_index_with, scope, Scope,
 };
 
 // ---------------------------------------------------------------------------